+++ packages/web3-core/src/utils.ts
@@ -23,11 +23,20 @@ import {
 	SupportedProviders,
 	Web3APISpec,
 	Web3BaseProvider,
+	MetaMaskProvider,
 } from 'web3-types';
 
 export const isWeb3Provider = <API extends Web3APISpec>(
 	provider: SupportedProviders<API>,
 ): provider is Web3BaseProvider<API> => Web3BaseProvider.isWeb3Provider(provider);
+
+export const isMetaMaskProvider = <API extends Web3APISpec>(
+	provider: SupportedProviders<API>,
+): provider is MetaMaskProvider<API> =>
+ typeof provider !== 'string' &&
+'request' in provider &&
+provider.request.constructor.name === 'AsyncFunction' && 'isMetaMask' in provider && provider.isMetaMask;
+
 
 export const isLegacyRequestProvider = <API extends Web3APISpec>(
 	provider: SupportedProviders<API>,
@@ -53,14 +62,13 @@ export const isLegacySendAsyncProvider = <API extends Web3APISpec>(
 
 export const isSupportedProvider = <API extends Web3APISpec>(
 	provider: SupportedProviders<API>,
-): provider is SupportedProviders<API> =>
+): provider is SupportedProviders<API> => 
 	provider &&
 	(isWeb3Provider(provider) ||
 		isEIP1193Provider(provider) ||
 		isLegacyRequestProvider(provider) ||
 		isLegacySendAsyncProvider(provider) ||
 		isLegacySendProvider(provider));
-
 export const isSupportSubscriptions = <API extends Web3APISpec>(
 	provider: SupportedProviders<API>,
 ): boolean => {