+++ packages/web3-utils/src/validation.ts
@@ -30,7 +30,7 @@ import {
 	isTopicInBloom as isTopicInBloomValidator,
 	isUserEthereumAddressInBloom as isUserEthereumAddressInBloomValidator,
 } from 'web3-validator';
-import { BlockNumberOrTag, BlockTags } from 'web3-types';
+import { BlockNumberOrTag, BlockTags, ContractInitOptions } from 'web3-types';
 
 /**
  * @deprecated Will be removed in next release. Please use `web3-validator` package instead.
@@ -168,5 +168,23 @@ export const compareBlockNumbers = (blockA: BlockNumberOrTag, blockB: BlockNumbe
 	}
 	return 1;
 };
+
+
+export const isContractInitOptions = (options: unknown): options is ContractInitOptions =>
+	typeof options === 'object' &&
+	!isNullishValidator(options) &&
+	Object.keys(options).length !== 0 &&
+	[
+		'input',
+		'data',
+		'from',
+		'gas',
+		'gasPrice',
+		'gasLimit',
+		'address',
+		'jsonInterface',
+		'syncWithContext',
+		'dataInputFill',
+	].some(key => key in options);
 
 export const isNullish = isNullishValidator;