+++ packages/web3-eth-contract/CHANGELOG.md
@@ -314,4 +314,16 @@ Documentation:
 
 -   The `events` property was added to the `receipt` object (#6410)
 
+## [4.1.2]
+
+### Changed
+
+-   Dependencies updated
+
+## [4.1.3]
+
+### Fixed
+
+-	Will populate `data` for transactions in contract for metamask provider instead of `input` (#6534)
+
 ## [Unreleased]