+++ packages/web3/CHANGELOG.md
@@ -161,4 +161,16 @@ Documentation:
 
 -   Various web3 sub packages has new functions details are in root changelog
 
+## [4.2.1]
+
+### Changed
+
+-   Dependencies updated
+
+## [4.2.2]
+
+### Changed
+
+-   Dependencies updated ( details are in root changelog )
+
 ## [Unreleased]